hw/vcpu_hp: new VCPU Hotplug device
Add the VCPU Hotplug Device, which is an arch-independent device which
can be used to construct and fire VCPU hotplug requests to the guest
via a binary MMIO interface. An interrupt is used to signal the guest
that it should read a new request. A control register set is used
afterwards to communicate between host and guest.

In the end, the guest clears the Hotplug pending register to signal
completion to this device. The vcpu_hp device in turn informs the vl
module that a hotplug completion event needs to be processed.

For now the hotplug completion event is just a no-operation.

Signed-off-by: Claudio Fontana <claudio.fontana@huawei.com>